+++ python/semantic-kernel-sampler-py/src/semantic_kernel_sampler/ai/modules/mcp_demo_server/sk/plugin.py
@@ -11,4 +11,4 @@ class DemoServerMCPStdioPlugin(MyMCPStdioPlugin):
 
     command: str = field(default="npm")
 
-    args: list[str] = field(default_factory=lambda: ["start", "--prefix", "../../node/modelcontextprotocol-typescript-sdk-example"])
+    args: list[str] = field(default_factory=lambda: ["start", "--prefix", "../../node/mcp-server.examples.getting-started"])